requirements.txt
requests
Flask
Flask-Cors
bcrypt
PyJWT
SQLAlchemy
pylint
python-dotenv
Flask-WTF
pymongo
